fix: correct wallet transaction precision and sell order status in ledger-service
WalletTransaction.amount truncated BigDecimal trade prices to Long via
longValue(), silently dropping fractional cents from the ledger history
even though the wallet balance itself was tracked correctly. Also,
sellAsset never marked the order SUCCESS after settling funds, leaving
every completed sell stuck at PENDING (buyAsset already did this
correctly).

+++ backend/ledger-service/src/main/java/dev/pioruocco/controller/WithdrawalController.java
@@ -12,6 +12,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 
@@ -58,7 +59,7 @@ public class WithdrawalController {
                 userWallet,
                 WalletTransactionType.WITHDRAWAL, null,
                 "bank account withdrawal",
-                withdrawal.getAmount()
+                BigDecimal.valueOf(withdrawal.getAmount())
         );
 
         return new ResponseEntity<>(withdrawal, HttpStatus.OK);
