+++ backend/ledger-service/src/main/java/dev/pioruocco/service/WalletTransactionServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -24,7 +25,7 @@ public class WalletTransactionServiceImpl implements WalletTransactionService {
                                                WalletTransactionType type,
                                                String transferId,
                                                String purpose,
-                                               Long amount
+                                               BigDecimal amount
     ) {
         WalletTransaction transaction = new WalletTransaction();
         transaction.setWallet(wallet);
